+++ pkg/interpreter/interpreter_integration_test.go
@@ -286,6 +286,63 @@ var integrationTests = []struct {
 		wantCategory:   "",
 		config:         interpreter.DefaultConfig(),
 	},
+	// v0.22.0 regression tests
+	{
+		name:           "atomic counter",
+		dir:            "atomic_counter",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "atomic cas",
+		dir:            "atomic_cas",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "io readall",
+		dir:            "io_readall",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "bufio scanner",
+		dir:            "bufio_scanner",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "strings builder",
+		dir:            "strings_builder",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "bytes buffer",
+		dir:            "bytes_buffer",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "log print",
+		dir:            "log_print",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "log fatal",
+		dir:            "log_fatal",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
 	// v0.21.0 regression tests
 	{
 		name:           "string byte index",